requirements/local.txt
-r base.txt

django-debug-toolbar==1.7
django-extensions==1.7.9

numpy==1.14.0
pandas==0.22.0
